refacotr : 공용 기능 global package로 재배치
