+++ src/main/java/kr/inuappcenterportal/inuportal/domain/category/service/CategoryService.java
@@ -3,8 +3,8 @@ package kr.inuappcenterportal.inuportal.domain.category.service;
 import kr.inuappcenterportal.inuportal.domain.category.model.Category;
 import kr.inuappcenterportal.inuportal.domain.category.dto.CategoryDto;
 import kr.inuappcenterportal.inuportal.domain.category.dto.CategoryUpdateDto;
-import kr.inuappcenterportal.inuportal.exception.ex.MyErrorCode;
-import kr.inuappcenterportal.inuportal.exception.ex.MyException;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyErrorCode;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import kr.inuappcenterportal.inuportal.domain.category.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
