+++ src/main/java/kr/inuappcenterportal/inuportal/domain/fire/service/FireService.java
@@ -8,8 +8,8 @@ import kr.inuappcenterportal.inuportal.domain.fire.dto.FireListResponseDto;
 import kr.inuappcenterportal.inuportal.domain.fire.dto.FirePageResponseDto;
 import kr.inuappcenterportal.inuportal.domain.fire.dto.FireRatingDto;
 import kr.inuappcenterportal.inuportal.domain.fire.dto.FireResponseDto;
-import kr.inuappcenterportal.inuportal.exception.ex.MyErrorCode;
-import kr.inuappcenterportal.inuportal.exception.ex.MyException;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyErrorCode;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import kr.inuappcenterportal.inuportal.domain.fire.repository.FireRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
