+++ src/main/java/kr/inuappcenterportal/inuportal/domain/post/service/PostService.java
@@ -13,8 +13,8 @@ import kr.inuappcenterportal.inuportal.dto.ListResponseDto;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostDto;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostListResponseDto;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostResponseDto;
-import kr.inuappcenterportal.inuportal.exception.ex.MyErrorCode;
-import kr.inuappcenterportal.inuportal.exception.ex.MyException;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyErrorCode;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import kr.inuappcenterportal.inuportal.service.RedisService;
 import kr.inuappcenterportal.inuportal.domain.reply.service.ReplyService;
 import lombok.RequiredArgsConstructor;
