--- src/main/java/kr/inuappcenterportal/inuportal/exception/CustomAuthenticationEntryPoint.java
@@ -1,27 +1,0 @@
-package kr.inuappcenterportal.inuportal.exception;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import kr.inuappcenterportal.inuportal.dto.ResponseDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        response.setContentType("application/json;charset=UTF-8");
-        String result = objectMapper.writeValueAsString(ResponseDto.of(401,"인증이 실패하였습니다."));
-        response.setStatus(401);
-        response.getWriter().write(result);
-    }
-}
