+++ src/main/java/kr/inuappcenterportal/inuportal/global/exception/CustomAccessDeniedHandler.java
@@ -1,0 +1,27 @@
+package kr.inuappcenterportal.inuportal.global.exception;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import kr.inuappcenterportal.inuportal.dto.ResponseDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+        response.setContentType("application/json;charset=UTF-8");
+        String result = objectMapper.writeValueAsString(ResponseDto.of(403,"접근 권한이 없는 사용자입니다."));
+        response.setStatus(403);
+        response.getWriter().write(result);
+    }
+}
