+++ src/main/java/kr/inuappcenterportal/inuportal/global/exception/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.exception;
+package kr.inuappcenterportal.inuportal.global.exception;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
@@ -6,22 +6,22 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import kr.inuappcenterportal.inuportal.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
 @Component
 @RequiredArgsConstructor
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
     private final ObjectMapper objectMapper;
 
     @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
         response.setContentType("application/json;charset=UTF-8");
-        String result = objectMapper.writeValueAsString(ResponseDto.of(403,"접근 권한이 없는 사용자입니다."));
-        response.setStatus(403);
+        String result = objectMapper.writeValueAsString(ResponseDto.of(401,"인증이 실패하였습니다."));
+        response.setStatus(401);
         response.getWriter().write(result);
     }
 }
