+++ src/main/java/kr/inuappcenterportal/inuportal/global/oracleRepository/SchoolLoginRepository.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.oracleRepository;
+package kr.inuappcenterportal.inuportal.global.oracleRepository;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
