+++ src/main/java/kr/inuappcenterportal/inuportal/service/RedisService.java
@@ -1,8 +1,8 @@
 package kr.inuappcenterportal.inuportal.service;
 
 import jakarta.xml.bind.DatatypeConverter;
-import kr.inuappcenterportal.inuportal.exception.ex.MyErrorCode;
-import kr.inuappcenterportal.inuportal.exception.ex.MyException;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyErrorCode;
+import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
